feat: converting encoders functions lib to typescript
